+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v1.2.0
+
+ - Add /unitinfo [`code`] command
+   - Search for any unit at QUT and recieve information about the unit including Prerequisites and Equivalents
+ - Do a little bit of code cleanup
+## v1.1.1
+
+ - update discord.js to fix breaking changes
+
 ## v1.1.0
 
 - Convert codebase to ESM modules